chore: move TextInput to common dir
